requirements.txt
biopython>=1.83
py-avataaars>=1.2
pillow>=10.0